# Lighthouse handler optimizations